Box: move TS types to types.ts
